+++ src/main/java/ar/edu/unq/dapp_api/model/builders/CryptoCurrencyBuilder.java
@@ -3,11 +3,12 @@ package ar.edu.unq.dapp_api.model.builders;
 import ar.edu.unq.dapp_api.model.CryptoCurrency;
 import ar.edu.unq.dapp_api.model.enums.CryptoSymbol;
 
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
 public class CryptoCurrencyBuilder {
     private CryptoSymbol symbol = CryptoSymbol.BTCUSDT;
-    private float price = 50000.0f;
+    private BigDecimal price = BigDecimal.valueOf(50000.0);
     private LocalTime lastUpdateDateAndTime = LocalTime.of(12, 0);
 
     public CryptoCurrencyBuilder withSymbol(CryptoSymbol symbol) {
@@ -15,7 +16,7 @@ public class CryptoCurrencyBuilder {
         return this;
     }
 
-    public CryptoCurrencyBuilder withPrice(float price) {
+    public CryptoCurrencyBuilder withPrice(BigDecimal price) {
         this.price = price;
         return this;
     }
