+++ src/main/java/ar/edu/unq/dapp_api/model/builders/OperationIntentBuilder.java
@@ -5,11 +5,13 @@ import ar.edu.unq.dapp_api.model.User;
 import ar.edu.unq.dapp_api.model.enums.CryptoSymbol;
 import ar.edu.unq.dapp_api.model.enums.IntentionType;
 
+import java.math.BigDecimal;
+
 public class OperationIntentBuilder {
     private CryptoSymbol symbol = CryptoSymbol.BTCUSDT;
-    private Long cryptoAmount = 1L;
-    private Float cryptoPrice = 50000F;
-    private Long operationARSAmount = 50000L;
+    private BigDecimal cryptoAmount = BigDecimal.valueOf(1);
+    private BigDecimal cryptoPrice = BigDecimal.valueOf(50000);
+    private BigDecimal operationARSAmount = BigDecimal.valueOf(50000);
     private User user = new UserBuilder().withId(1L).build();
     private IntentionType type = IntentionType.BUY;
 
@@ -18,17 +20,17 @@ public class OperationIntentBuilder {
         return this;
     }
 
-    public OperationIntentBuilder withCryptoAmount(Long cryptoAmount) {
+    public OperationIntentBuilder withCryptoAmount(BigDecimal cryptoAmount) {
         this.cryptoAmount = cryptoAmount;
         return this;
     }
 
-    public OperationIntentBuilder withCryptoPrice(Float cryptoPrice) {
+    public OperationIntentBuilder withCryptoPrice(BigDecimal cryptoPrice) {
         this.cryptoPrice = cryptoPrice;
         return this;
     }
 
-    public OperationIntentBuilder withOperationARSAmount(Long operationARSAmount) {
+    public OperationIntentBuilder withOperationARSAmount(BigDecimal operationARSAmount) {
         this.operationARSAmount = operationARSAmount;
         return this;
     }
