+++ src/main/java/ar/edu/unq/dapp_api/webservice/dto/operationIntent/NewOperationIntentDTO.java
@@ -4,14 +4,16 @@ import ar.edu.unq.dapp_api.model.enums.CryptoSymbol;
 import ar.edu.unq.dapp_api.model.enums.IntentionType;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class NewOperationIntentDTO {
 
     private CryptoSymbol symbol;
-    private Long cryptoAmount;
+    private BigDecimal cryptoAmount;
     private IntentionType type;
 
-    public NewOperationIntentDTO(CryptoSymbol symbol, Long cryptoAmount, IntentionType type) {
+    public NewOperationIntentDTO(CryptoSymbol symbol, BigDecimal cryptoAmount, IntentionType type) {
         this.symbol = symbol;
         this.cryptoAmount = cryptoAmount;
         this.type = type;
